passing in windowId to fix shotcut key issue. wasn't working with only focus param. comment out still broken areas ga analytics, github star widget
