+++ options.html
@@ -6,7 +6,7 @@
 	</head>
 		<h1>Tab Resize - Options</h1>
 		<br/>
-		<iframe class="star-gazer" src="https://ghbtns.com/github-btn.html?user=peterdotjs&repo=tab-resize&type=star&count=true&size=large" frameborder="0" scrolling="0" width="160px" height="30px"></iframe>
+		<!-- <iframe class="star-gazer" src="https://ghbtns.com/github-btn.html?user=peterdotjs&repo=tab-resize&type=star&count=true&size=large" frameborder="0" scrolling="0" width="160px" height="30px"></iframe> -->
 
 		<section id="tips">
 			<h2>Tips</h2>
